requirements.txt
aniso8601==9.0.1
autopep8==1.5.6
click==7.1.2
Flask==1.1.2
Flask-RESTful==0.3.8
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
mypy==0.812
mypy-extensions==0.4.3
psycopg2-binary==2.8.6
pycodestyle==2.7.0
pytz==2021.1
rope==0.18.0
six==1.15.0
toml==0.10.2
typed-ast==1.4.3
typing-extensions==3.7.4.3
Werkzeug==1.0.1